+++ Hikari Engine/src/Vector.cpp
@@ -4,31 +4,31 @@ Hikari::Vector2D::Vector2D() : m_x(0), m_y(0)
 {
 }
 
-Hikari::Vector2D::Vector2D(double x, double y): m_x(x), m_y(y)
+Hikari::Vector2D::Vector2D(float x, float y): m_x(x), m_y(y)
 {
 }
 
-double Hikari::Vector2D::x(void) const
+float Hikari::Vector2D::x(void) const
 {
 	 return m_x;
 }
 
-double Hikari::Vector2D::y(void) const
+float Hikari::Vector2D::y(void) const
 {
 	 return m_y;
 }
 
-void Hikari::Vector2D::x(double x)
+void Hikari::Vector2D::x(float x)
 {
 	m_x = x;
 }
 
-void Hikari::Vector2D::y(double y)
+void Hikari::Vector2D::y(float y)
 {
 	m_y = y;
 }
 
-Hikari::Vector3D::Vector3D(double x, double y, double z) : Vector2D(x, y), m_z(z)
+Hikari::Vector3D::Vector3D(float x, float y, float z) : Vector2D(x, y), m_z(z)
 {
 }
 
@@ -40,29 +40,29 @@ Hikari::Vector3D::Vector3D(const Vector3D& start, const Vector3D& end) : Vector2
 {
 }
 
-double Hikari::Vector3D::z(void) const
+float Hikari::Vector3D::z(void) const
 {
 	 return m_z;
 }
 
-void Hikari::Vector3D::z(double z)
+void Hikari::Vector3D::z(float z)
 {
 	m_z = z;
 }
 
-double Hikari::Vector3D::length(void) const
+float Hikari::Vector3D::length(void) const
 {
 	return sqrt(m_x * m_x + m_y * m_y + m_z * m_z);
 }
 
-double Hikari::Vector3D::lengthSqrd(void) const
+float Hikari::Vector3D::lengthSqrd(void) const
 {
 	return m_x * m_x + m_y * m_y + m_z * m_z;
 }
 
 void Hikari::Vector3D::normalize(void)
 {
-	double len = length();
+	float len = length();
 	m_x /= len;
 	m_y /= len;
 	m_z /= len;
@@ -70,7 +70,7 @@ void Hikari::Vector3D::normalize(void)
 
 Hikari::Vector3D Hikari::Vector3D::normalized(void)
 {
-	double len = length();
+	float len = length();
 	return Vector3D(m_x/len, m_y/len, m_z/len);
 }
 
@@ -92,11 +92,11 @@ Hikari::Vector3D Hikari::operator- (const Hikari::Vector3D& rFirst, const Hikari
 							);
 }
 
-Hikari::Vector4D::Vector4D(double x, double y, double z, double w) : Vector3D(x, y, z), m_w(w)
+Hikari::Vector4D::Vector4D(float x, float y, float z, float w) : Vector3D(x, y, z), m_w(w)
 {
 }
 
-Hikari::Vector4D::Vector4D(const Vector3D& rVector, double w) : Vector3D(rVector), m_w(w)
+Hikari::Vector4D::Vector4D(const Vector3D& rVector, float w) : Vector3D(rVector), m_w(w)
 {
 }
 
@@ -104,12 +104,12 @@ Hikari::Vector4D::Vector4D(const Vector3D& rStart, const Vector3D& rEnd) : Vecto
 {
 } // czy na pewno 0 to wektor?
 
-double Hikari::Vector4D::w(void) const
+float Hikari::Vector4D::w(void) const
 {
 	return m_w;
 }
 
-void Hikari::Vector4D::w(double w)
+void Hikari::Vector4D::w(float w)
 {
 	m_w = w;
 }
